requirements.txt
torch==2.3.0
transformers
pyyaml
python-dotenv
numpy
requests